support db mapping file for snowflake create table action

+++ components/components-snowflake/components-snowflake-runtime/src/main/java/org/talend/components/snowflake/runtime/tableaction/SnowflakeTableActionConfig.java
@@ -43,8 +43,8 @@ public class SnowflakeTableActionConfig extends TableActionConfig {
         //where to use this? not adjust now as decimal(32, 0) is ok for float, seems no data loss
         this.CONVERT_LOGICALTYPE_TO_SQLTYPE.put(LogicalTypes.decimal(32), Types.FLOAT);
 
-        //before we map to Types.FLOAT, not right as data loss
-        this.CONVERT_JAVATYPE_TO_SQLTYPE.put("java.math.BigDecimal", Types.DECIMAL);
+        //we map to Types.FLOAT, not right as data loss, TODO correct it
+        this.CONVERT_JAVATYPE_TO_SQLTYPE.put("java.math.BigDecimal", Types.FLOAT);
 
         this.CONVERT_SQLTYPE_TO_ANOTHER_SQLTYPE.put(Types.BLOB, Types.BINARY);
